# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Espaço Bíblico Infantil API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  auxiliary: 'auxiliary',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  congregationId?: number | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Congregation {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  responsibleName?: string | null;
  createdAt?: string;
}

export interface CongregationInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  state?: string;
  responsibleName?: string;
}

export interface CongregationUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  city?: string;
  state?: string;
  responsibleName?: string;
}

export type ChildGender = typeof ChildGender[keyof typeof ChildGender];


export const ChildGender = {
  M: 'M',
  F: 'F',
} as const;

export interface Child {
  id: number;
  fullName: string;
  birthDate: string;
  gender: ChildGender;
  guardianName: string;
  guardianPhone: string;
  foodRestriction?: boolean;
  /** @nullable */
  foodRestrictionDescription?: string | null;
  autism?: boolean;
  /** @nullable */
  observations?: string | null;
  /** @nullable */
  congregationId?: number | null;
  /** @nullable */
  congregationName?: string | null;
  age?: number;
  presentToday?: boolean;
  createdAt?: string;
}

export type ChildInputGender = typeof ChildInputGender[keyof typeof ChildInputGender];


export const ChildInputGender = {
  M: 'M',
  F: 'F',
} as const;

export interface ChildInput {
  /** @minLength 1 */
  fullName: string;
  birthDate: string;
  gender: ChildInputGender;
  /** @minLength 1 */
  guardianName: string;
  /** @minLength 1 */
  guardianPhone: string;
  foodRestriction?: boolean;
  foodRestrictionDescription?: string;
  autism?: boolean;
  observations?: string;
  congregationId?: number;
}

export type ChildUpdateGender = typeof ChildUpdateGender[keyof typeof ChildUpdateGender];


export const ChildUpdateGender = {
  M: 'M',
  F: 'F',
} as const;

export interface ChildUpdate {
  /** @minLength 1 */
  fullName?: string;
  birthDate?: string;
  gender?: ChildUpdateGender;
  guardianName?: string;
  guardianPhone?: string;
  foodRestriction?: boolean;
  foodRestrictionDescription?: string;
  autism?: boolean;
  observations?: string;
  congregationId?: number;
}

export type ChildWithFrequencyGender = typeof ChildWithFrequencyGender[keyof typeof ChildWithFrequencyGender];


export const ChildWithFrequencyGender = {
  M: 'M',
  F: 'F',
} as const;

export interface ChildWithFrequency {
  id: number;
  fullName: string;
  birthDate: string;
  gender: ChildWithFrequencyGender;
  guardianName: string;
  guardianPhone: string;
  foodRestriction?: boolean;
  /** @nullable */
  foodRestrictionDescription?: string | null;
  autism?: boolean;
  /** @nullable */
  observations?: string | null;
  /** @nullable */
  congregationId?: number | null;
  /** @nullable */
  congregationName?: string | null;
  age?: number;
  presentToday?: boolean;
  createdAt?: string;
  absenceCount: number;
  presenceCount?: number;
  /** @nullable */
  lastPresence?: string | null;
}

export interface ChildFrequency {
  childId: number;
  presenceCount: number;
  absenceCount: number;
  frequencyPercent: number;
  /** @nullable */
  lastPresence?: string | null;
}

export interface Attendance {
  id: number;
  childId: number;
  attendanceDate: string;
  attendanceTime: string;
}

export interface AttendanceInput {
  childId: number;
  attendanceDate?: string;
}

export interface AttendanceWithChild {
  id: number;
  childId: number;
  attendanceDate: string;
  attendanceTime: string;
  child: Child;
}

export interface DailyReport {
  /** @nullable */
  id?: number | null;
  reportDate: string;
  totalGirls: number;
  totalBoys: number;
  totalChildren: number;
}

export interface DashboardStats {
  totalChildren: number;
  totalGirls: number;
  totalBoys: number;
  presentToday: number;
  absentToday: number;
  todayGirls: number;
  todayBoys: number;
  weeklyAverage?: number;
}

export interface WeeklyStat {
  weekLabel: string;
  reportDate?: string;
  totalChildren: number;
  totalGirls: number;
  totalBoys: number;
}

export type ListChildrenParams = {
/**
 * Search by name
 */
search?: string;
gender?: ListChildrenGender;
congregationId?: number;
autism?: boolean;
foodRestriction?: boolean;
ageAbove12?: boolean;
presentToday?: boolean;
absentToday?: boolean;
};

export type ListChildrenGender = typeof ListChildrenGender[keyof typeof ListChildrenGender];


export const ListChildrenGender = {
  M: 'M',
  F: 'F',
} as const;

export type GetMostAbsentChildrenParams = {
limit?: number;
};

export type ListAttendanceParams = {
/**
 * Filter by date (YYYY-MM-DD)
 */
date?: string;
};

export type ListDailyReportsParams = {
startDate?: string;
endDate?: string;
};