# lib/api-zod/src/generated/types/childFrequency.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Espaço Bíblico Infantil API
 * OpenAPI spec version: 0.1.0
 */

export interface ChildFrequency {
  childId: number;
  presenceCount: number;
  absenceCount: number;
  frequencyPercent: number;
  /** @nullable */
  lastPresence?: string | null;
}